refactor: changed pods structure
