+++ src/account/index.ts
@@ -1,4 +1,4 @@
 export * as Account from './account'
-export * as Encryption from './encryption'
+export * as Encryption from '../utils/encryption'
 export * as Utils from './utils'
 //TODO export every exportable
