+++ src/pod/list.ts
@@ -1,4 +1,5 @@
 import { Pod, SharedPod } from './types'
+import { assertPodList, assertSharedPodList } from './utils'
 
 /**
  * List of created and shared pods
@@ -18,5 +19,18 @@ export class List {
    */
   getSharedPods(): SharedPod[] {
     return this.sharedPods
+  }
+
+  /**
+   * Creates pods list from JSON string
+   *
+   * @param json {pods: Pod[], sharedPods: SharedPod[]}
+   */
+  static fromJSON(json: string): List {
+    const object = JSON.parse(json)
+    assertPodList(object.pods)
+    assertSharedPodList(object.sharedPods)
+
+    return new List(object.pods, object.sharedPods)
   }
 }
